+++ src/core/command/animation_commands.hpp
@@ -211,11 +211,15 @@ public:
     void undo() override
     {
         target->stretch_time(1/multiplier);
+        if constexpr ( !std::is_same_v<T, model::Document> )
+            target->set_time(target->document()->current_time());
     }
 
     void redo() override
     {
         target->stretch_time(multiplier);
+        if constexpr ( !std::is_same_v<T, model::Document> )
+            target->set_time(target->document()->current_time());
     }
 
 private: